requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.9
sqlalchemy-utils>=0.40.0
alembic>=1.10.3
pydantic[email]>=1.10.7
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
pytest>=7.3.1
httpx>=0.24.0
bcrypt>=4.0.1
email-validator>=2.0.0
